aout: Add a 'mono' option to the stereo-mode parameter
This parameter value needs to be dealt with before starting the
actual aout, which is why the aout->Start call is moved down
to after the stereo-mode parameter is created.

---
Tested to work properly with the android_audiotrack output
(with a patch that adds a gui preference for enabling this mode).

Also tested on linux with the qt4 gui, where the new option appeared
properly in the gui and seemed to work as intended.

(However, the stereo mode submenu appeared grayed out when playing an
8-channel audio file, while it appeared normally and worked fine when
playing a video with 2 audio channels.)