omxil/mediacodec: Use the top/left crop parameters as offset
The only off-hand known codec where these are nonzero are in the
ducati codecs on the OMAP4. The offset buffer parameter also includes
these values from the start, so we can't respect both the crop offset
and the buffer offset at the same time.

This is more in line with what google's CTS tests for handling
raw YUV data in MediaCodec in Android 4.3 shows.

For MediaCodec, just ignore the offset parameter outright. This
is what the Android CTS test does anyway. For "normal" omxil,
just ignore it for the OMAP4 specific pixel format.