Don't build glew32.dll at all
The previous win32.patch removes some flags from the build of glew32.dll,
so that it tries to link glew32.dll without -shared, effectively
producing an exe instead of a dll.

Additionally, we can't use $(LD) directly for linking (as the makefile
does), since that will invoke the ELF side of LLD (unless the caller
provides flags that indicate that it is for a PE/COFF target. By
using $(CC) instead of $(LD), linking of the DLL could have a chance
of working, but it won't since win32.patch removed the necessary -shared
flag.

Since the DLLs aren't desired at all, just remove them.