Patch gme to add some dllexport declarations at the start of a source file
Clang considers it incorrect to add more dll attributes after the
declaration has been seen. Normally it only warns about this, but
it is considered an error if the declaration has been used before.
Therefore, functions that are referenced before their definition
within gme.cpp will be considered an error.

The public header gme.h doesn't include blargg_source.h that provides
the dllexport macro, but we can declare the problematic functions
at the top of the source file with the right attributes.