Clean libQt5Bootstrap.a in the build tree as well
When building the Qt tools for the target architecture in cross
compilation, the bootstrap object library is rebuilt, but the
actual static library itself wasn't removed before. This meant
that the object files were added to/replacing the previous object
files (for the build host). Object files that were produced in the
host build but not in the cross build were left over.

lld-link errors out when it encounters object files in a static library
that it can't parse (ELF files when doing a mingw cross build).